Added DynamicMapIcon

+++ src/SampSharp.Streamer/Natives/StreamerNative.map-icons.cs
@@ -11,6 +11,7 @@
 // 
 // For more information, please refer to <http://unlicense.org>
 
+using System;
 using SampSharp.GameMode.Definitions;
 using SampSharp.GameMode.Natives;
 using SampSharp.GameMode.SAMP;
@@ -23,7 +24,7 @@ namespace SampSharp.Streamer.Natives
             int interiorid = -1, int playerid = -1, float streamdistance = 100.0f, MapIconType style = MapIconType.Local)
         {
             return Native.CallNative("CreateDynamicMapIcon",
-                __arglist(x, y, z, type, (int) color, worldid, interiorid, playerid, streamdistance, style));
+                __arglist(x, y, z, type, (int) color, worldid, interiorid, playerid, streamdistance, (int)style));
         }
